requirements.txt
certifi==2024.8.30
charset-normalizer==3.4.0
idna==3.10
MouseInfo==0.1.3
PyAutoGUI==0.9.54
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyobjc-core==10.3.1
pyobjc-framework-Cocoa==10.3.1
pyobjc-framework-Quartz==10.3.1
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==1.0.1
pyTelegramBotAPI==4.23.0
python-dotenv==1.0.1
pytweening==1.2.0
requests==2.32.3
rubicon-objc==0.4.9
urllib3==2.2.3
